llm/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
groq
pandas
numpy
faiss-cpu
tqdm
sentence-transformers